requirements.txt
# Core Dependencies
flask>=2.0.1
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.2
tensorflow>=2.7.0
keras>=2.7.0

# Data Sources & APIs
python-binance>=1.0.15
pycoingecko>=2.2.0
glassnode-sdk>=0.1.0
newsapi-python>=0.2.6
python-twitter>=3.5.0
praw>=7.4.0

# Advanced Analytics
xgboost>=1.5.0
lightgbm>=3.3.2
arch>=4.19.0  # For GARCH models
statsmodels>=0.13.2
prophet>=1.0.1
isolationforest>=0.1.0

# Natural Language Processing
nltk>=3.6.5
textblob>=0.15.3
transformers>=4.12.0

# Visualization
plotly>=5.3.1
dash>=2.0.0
dash-bootstrap-components>=1.0.0
matplotlib>=3.4.3
seaborn>=0.11.2

# Database
pymongo>=3.12.0
psycopg2-binary>=2.9.1
sqlalchemy>=1.4.23

# Utils
python-dotenv>=0.19.0
requests>=2.26.0
websocket-client>=1.2.1
ccxt>=1.60.0  # For crypto exchange data
